+++ internal/routesync/route.go
@@ -333,3 +333,18 @@ func (s *Sync) routeDownAction(m *rtnetlink.RouteMessage) error {
 	m.Attributes.Table = s.table
 	return s.nlconn.Route.Delete(m)
 }
+
+func ChangeMetric(conn *rtnetlink.Conn, msg rtnetlink.RouteMessage, metric uint32) error {
+	orgMetrc := msg.Attributes.Priority
+	msg.Flags = 0
+	// we add first, to prevent moment without route
+	msg.Attributes.Priority = metric
+	if err := conn.Route.Add(&msg); err != nil {
+		return fmt.Errorf("change metric error: unable to add route: %w", err)
+	}
+	msg.Attributes.Priority = orgMetrc
+	if err := conn.Route.Delete(&msg); err != nil {
+		return fmt.Errorf("change metric error: unable to delete route: %w", err)
+	}
+	return nil
+}
